fix type-check false-positives and new tests

+++ test/unusedvar.jl
@@ -30,7 +30,7 @@ msgs = lintstr(s)
 @test( contains( msgs[1].message, "declared but not used" ))
 s = """
 function f(x)
-    local a
+    local a::Int
     local b = 2::Int # type assertion/conversion
     return x+b
 end
